services/embed/requirements.txt
# ============================================================================
# FILE: services/embed/requirements.txt
# PURPOSE: Python dependencies for the embedding service.
# ARCHITECTURE REF: §3.3 — BGE-M3 Optimization for CPU
# ============================================================================

# FastAPI framework + ASGI server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Pydantic settings for environment variable configuration
pydantic-settings==2.3.0

# FlagEmbedding: Official library for BGE models (dense + sparse in one pass)
# Supports BGE-M3's hybrid retrieval (ColBERT dense + BM25 sparse)
FlagEmbedding==1.2.11
transformers==4.38.2
peft==0.10.0               # Pinned to avoid 'Optional' bug in newer versions
numpy<2                    # Pinned to fix NumPy 2.x C-extension compatibility bug
# PyTorch CPU-only (saves ~1.5 GB vs CUDA version)
# Install via --index-url https://download.pytorch.org/whl/cpu
# Pinned version for reproducibility
torch==2.2.2+cpu; platform_machine=="x86_64"
torch==2.2.2; platform_machine=="aarch64"

# Prometheus metrics instrumentation for FastAPI
prometheus-fastapi-instrumentator==6.1.0

# HTTP client (for health checks to dependencies if needed)
httpx==0.27.0

# tenacity: retry library with exponential backoff
tenacity==8.3.0
